added routes with react-router-dom
