+++ package.json
@@ -9,10 +9,13 @@
     "bootstrap": "^5.3.0",
     "bootstrap-icons": "^1.10.5",
     "cors": "^2.8.5",
+    "date-fns": "^2.30.0",
     "pg": "^8.11.1",
     "react": "^18.2.0",
+    "react-big-calendar": "^1.8.1",
     "react-bootstrap": "^2.8.0",
     "react-bootstrap-icons": "^1.10.3",
+    "react-datepicker": "^4.15.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.14.0",
     "react-scripts": "5.0.1",
