+++ src/App.js
@@ -1,25 +1,19 @@
-import TopBar from "./pages/home-page/TopBar";
-import MidImgLogo from "./pages/home-page/logoWithBg";
-import InfoCards from "./pages/home-page/InfoCards";
-import Products from "./pages/home-page/Products";
-import Footer from "./pages/home-page/Footer";
-import Cal1 from "./componets/Calander";
 import "./App.css";
+import { BrowserRouter as Router, Route, Routes } from 'react-router-dom'
+import HomePage from "./pages/HomePage";
+import CalendarPage from "./pages/CalanderPage";
 
 function App() {
     return (
         <div className="App">
-            <body className="bg-rose-500">
-                <div className="TopBar">
-                    <TopBar />
-                </div>
-                <div className="max-w-7xl mx-auto ">
-                    <MidImgLogo />
-                    <InfoCards />
-                    <Products />
-                </div>
-                <Footer />
-            </body>
+            <Router>
+                <Routes>
+                    <Route path='/' element={<HomePage/>}/>
+                </Routes>
+                <Routes>
+                    <Route path='/Calendar' element={<CalendarPage/>}/>
+                </Routes>
+            </Router>
         </div>
     );
 }
