--- src/assets/images/test
@@ -1,1 +1,0 @@
-//fill in
