+++ src/componets/InfoCards.jsx
@@ -2,8 +2,8 @@
 import CardGroup from "react-bootstrap/CardGroup";
 import Card from "react-bootstrap/Card";
 import ListGroup from "react-bootstrap/ListGroup";
-import Gel from "../img/GelNail.jpg";
-import Acr from "../img/acrylicNail.jpg";
+import Gel from "../assets/images/GelNail.jpg";
+import Acr from "../assets/images/acrylicNail.jpg";
 
 //Function
 function InfoCards() {
