+++ src/pages/Error404.jsx
@@ -1,9 +1,9 @@
 import React from 'react'
 
-function Error404() {
+function Error404Page() {
   return (
     <div>Error404</div>
   )
 }
 
-export default Error404
+export default Error404Page
